+++ capytaine/bem/solver.py
@@ -23,6 +23,7 @@ from capytaine.bem.engines import BasicMatrixEngine
 from capytaine.io.xarray import problems_from_dataset, assemble_dataset, kochin_data_array
 from capytaine.tools.optional_imports import silently_import_optional_dependency
 from capytaine.tools.lists_of_points import _normalize_points, _normalize_free_surface_points
+from capytaine.tools.symbolic_multiplication import supporting_symbolic_multiplication
 
 LOG = logging.getLogger(__name__)
 
@@ -93,7 +94,8 @@ class BEMSolver:
             problem.free_surface, problem.water_depth, problem.wavenumber,
             self.green_function
         )
-        sources = self.engine.linear_solver(K, problem.boundary_condition)
+        linear_solver = supporting_symbolic_multiplication(self.engine.linear_solver)
+        sources = linear_solver(K, problem.boundary_condition)
         potential = S @ sources
         pressure = 1j * problem.omega * problem.rho * potential
